requirements.txt
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0